fix(complex numbers): fix return type for modulus method float -> double

+++ header/complex_numbers.hpp
@@ -14,7 +14,7 @@ public:
 
     Complex operator+(const Complex& other) const;
     Complex conjugate() const;
-    float modulus() const;
+    double modulus() const;
 
 
 };
